Initial commit
This is a working emulator for the Enigma cipher machine.

This version emulates the machine itself, but does not handle
looking up the key from file or figuring out which parts of a
message is metadata.

It does agree with a few online emulators and it does give the
plain text back if the enciphered text is enciphered again.