Half-transformed enigma.
This version is an attempt to write the enigma fully
automated eith an operator func/object to take care of
initial text cleaning, finding the correct key and figuring
out the metadata parts of the messages.

This version is not finished and does not run. Work is needed.

Some good work has been done and should be used when writing
the new and better enigma this was meant to be.