This version runs partly.
The convertion from functions to classes have just begun.
This version runs, but only in the encipher mode, not
in decipher mode. The key is read from file, the message
key is chosen at random and the cipher text is formated
correctly.

Next up is to get the decipher mode to run.