Operator handles arbitrary number of rotors.
Operator.encipher now pick starting positions for as many
rotors as specified by the key read from file. Earlier the
number of starting positions was hard coded to 3.

Operator.decipher now looks for message starts positions and
encoded message key which are not necessarily of length 3.

Due to the double stepping feature, the number of rotors may
not be less than 2. This restriction is not seen as a weakness,
as the point of this change is to increase the number of rotors.