Starting to implement extended alphabet.
Replacing the 26 character upper case only alphabet with a
233 character alphabet including upper and lower characters for
latin, nordic, greek and russian in addition to most common
punctuation.

Space (\u0020) is omitted for now due to the expected extra
effort needed to get things to work.

Current version does not run because there is no working
reflector with all characters.