Barely working version with extended alphabet.
Further extended the alphabet to include runes.

Added atbash reflector and the program runs and reproduces the
plain text when run both ways.

There is a problem with how the position of the rotors is handeled.
The position is a number which increases mod len alphabet.
Now, this creates problems because not all characters in the
Unicode table area used in the alphabet, thus getting positions
which should not be possible.